Real Estate Price Prediction Project
This is the Real Price Prediction Project of Python that contains Data Visualization and Machine Learning. 
In this project,  we have been given a dataset that describe the functionality of houses. Now based on the given features, we have to predict the house price.
This project uses the following libraries and modules of python:
1.Numpy
2.Pandas
3.Matplotlib
4.Seaborn
5.Sklearn
6.Re module