Pokemon_Visualization_Project
This is the Python and Data Visualization project. In this project,  we have been given a dataset that describe the functionality and features of different Pokémon's. Now based on the given information, we have to plot different graphs according to their features. 
This project uses the following libraries and modules of python: 
1.Numpy 
2.Pandas 
3.Matplotlib 
4.Seaborn